requirements.txt
# Основные зависимости
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# База данных и миграции
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
asyncpg==0.29.0

# Валидация и сериализация
pydantic==2.5.3
pydantic-settings==2.1.0

# Безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1

# Санитизация HTML
bleach==6.1.0

# Утилиты
python-slugify==8.0.4
python-multipart==0.0.6

# Тестирование
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Логирование
loguru==0.7.2

email-validator